% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get.deriv.params}
\alias{get.deriv.params}
\title{Get parameters to initialize numerical differentiation}
\usage{
get.deriv.params(EpsNumericalGrad = 1e-06, EpsNumericalHessEta = 1e-04)
}
\arguments{
\item{EpsNumericalGrad}{The "epsilon" used to perturb the objective functon
when computing numerical gradients}

\item{EpsNumericalHessEta}{The "epsilon" used to perturb the objective functon
when computing numerical HessEta}
}
\value{
List containing input arguments for numerical differentiation
}
\description{
Get parameters to initialize numerical differentiation
}
