% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get.manifold.params}
\alias{get.manifold.params}
\title{Get parameters to initialize manifold}
\usage{
get.manifold.params(IsCheckParams = FALSE)
}
\arguments{
\item{IsCheckParams}{Should internal manifold object check inputs and
print summary message before optimization (TRUE or FALSE)}
}
\value{
List containing input arguments for manifold
}
\description{
Get parameters to initialize manifold
}
