% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print.ManifoldOptim}
\alias{print.ManifoldOptim}
\title{Print summary from \code{manifold.optim} results}
\usage{
\method{print}{ManifoldOptim}(x, ...)
}
\arguments{
\item{x}{A \code{ManifoldOptim} object output by \code{manifold.optim}.}

\item{...}{Not currently used}
}
\description{
Print results
}
